+++ packages/core/src/types/send.ts
@@ -1,0 +1,10 @@
+import type { WalkerOS } from '.';
+
+export type SendDataValue = WalkerOS.Property | WalkerOS.Properties;
+export type SendHeaders = { [key: string]: string };
+
+export interface SendResponse {
+  ok: boolean;
+  data?: unknown;
+  error?: string;
+}
